requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
amqp==5.1.1
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
billiard==3.6.4.0
celery==5.2.6
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
crispy-bootstrap5==0.6
Deprecated==1.2.13
Django==4.0.4
django-appconf==1.0.5
django-compressor==4.0
django-crispy-forms==1.14.0
django-debug-toolbar==3.2.4
django-elasticsearch-dsl==7.2.2
django-elasticsearch-dsl-drf==0.22.4
django-extensions==3.1.5
django-filter==21.1
django-model-utils==4.2.0
django-nine==0.2.5
django-npm==1.0.0
django-redis==5.2.0
django-sass-processor==1.1
django-seed==0.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
elasticsearch==7.17.3
elasticsearch-dsl==7.4.0
Faker==13.6.0
frozenlist==1.3.0
idna==3.3
kombu==5.2.4
libsass==0.21.0
Markdown==3.3.6
multidict==6.0.2
mypy-extensions==0.4.3
mysql-connector-python==8.0.28
packaging==21.3
pathspec==0.9.0
Pillow==9.1.0
platformdirs==2.5.2
prompt-toolkit==3.0.29
psycopg2==2.9.3
PyJWT==1.7.1
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
rcssmin==1.1.0
redis==4.2.2
rjsmin==1.2.0
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
toposort==1.7
tzdata==2022.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.0
yarl==1.7.2
